delete Component DeleteColor, integrate the function in Color Component
