+++ src/Components/Color/Color.css
@@ -9,3 +9,8 @@
   background: black;
   color: white;
 }
+
+.deleteButton {
+  box-shadow: 5px 3px 5px #000;
+  margin-bottom: 10px;
+}
