--- src/Components/DeleteColor/DeleteColor.css
@@ -1,4 +1,0 @@
-.deleteButton {
-  box-shadow: 5px 3px 5px #000;
-  margin-bottom: 10px;
-}
